separate logic between chat and freechat & handle conversational logic from chat-with-llm
